Factor out widget classes to external file
